Remove `include/grpcpp/impl/codegen/client_interceptor.h`
